reorganized Ollama examples
